refactor(reflection): HLE-53 dissolve atman.reflection module into adapters
The atman.reflection/ module lived outside the core/ ↔ adapters/ layering
and contained two violations: store.py imported AgentSchemaResolver from
adapters/ (wrong direction), and models.py declared its own ReflectionLevel
enum duplicating atman.core.models.reflection.ReflectionLevel.

Move:
- reflection/store.py → adapters/storage/reflection_store.py
- reflection/models.py → adapters/storage/reflection_persistence_models.py
  (renamed to make the persistence-layer purpose explicit and distinguish
  from the domain model in core/models/reflection.py)

The persistence model now re-imports ReflectionLevel from core so the two
layers share a single source of truth. Update test imports + remove the
stale atman.reflection entry from .importlinter.
