--- src/atman/reflection/__init__.py
@@ -1,11 +1,0 @@
-"""
-Reflection persistence layer.
-
-This module provides persistent storage for reflection content,
-separate from the core domain models in atman.core.models.reflection.
-"""
-
-from atman.reflection.models import ReflectionEvent
-from atman.reflection.store import ReflectionStore
-
-__all__ = ["ReflectionEvent", "ReflectionStore"]
